[feat] ProductService - registerProduct() 완성

+++ src/main/java/com/domain/service/ProductService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ProductService {
     List<ProductDto> getAllProducts();
     ProductDto findProductByProductId(int productId);
-    int registerProduct(ProductDto productDto);
+    Long registerProduct(ProductDto productDto);
     ProductDto updateProductDetail(int productId, ProductDto productDto);
     int removeProduct(int productId);
     List<ProductDto> findProductsBySearchingKeyword(String keyword, int price, boolean sale);
